requirements.txt
arxiv==2.4.0
requests==2.32.5
pyyaml==6.0.3
schedule==1.2.2
playwright==1.57.0
playwright-stealth==2.0.0
PyPDF2==3.0.1
lxml==6.0.2
beautifulsoup4==4.14.3
feedparser==6.0.12
xhtml2pdf==0.2.17
